backend/requirements.txt
Flask==2.2.3
azure-functions==1.11.0
python-dotenv==0.21.0
requests==2.28.2
gunicorn==20.1.0
